requirements.txt
attrs==21.3.0
bootstrapped==0.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
Deprecated==1.2.13
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
fonttools==4.28.5
GitHub-Flask==3.2.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mongoengine==0.23.1
mpmath==1.2.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
parameterized==0.8.1
patsy==0.5.2
Pillow==8.4.0
pipfile==0.0.2
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
pymongo==3.12.3
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.29
statsmodels==0.13.1
sympy==1.9
tealang==0.4.0
threadpoolctl==3.0.0
toml==0.10.2
urllib3==1.26.7
uuid==1.30
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
z3-solver==4.8.14.0
